Moved utility from doe to parmest
