+++ pyomo/contrib/parmest/utils/model_utils.py
@@ -13,8 +13,8 @@ import logging
 
 import pyomo.environ as pyo
 from pyomo.core.expr import replace_expressions, identify_mutable_parameters
-from pyomo.core.base.var import IndexedVar
-from pyomo.core.base.param import IndexedParam
+from pyomo.core.base.var import IndexedVar, VarData
+from pyomo.core.base.param import IndexedParam, ParamData
 from pyomo.common.collections import ComponentMap
 
 from pyomo.environ import ComponentUID
@@ -201,3 +201,37 @@ def convert_params_to_vars(model, param_names=None, fix_vars=False):
     # solver.solve(model)
 
     return model
+
+
+# Adding utility to update parameter values in a model based on the suffix
+def update_model_from_suffix(suffix_obj: pyo.Suffix, values):
+    """
+    Overwrite each variable/parameter referenced by ``suffix_obj`` with the
+    corresponding value in ``values``.
+
+    Parameters
+    ----------
+    suffix_obj : pyomo.core.base.suffix.Suffix
+        The suffix whose *keys* are the components you want to update.
+        Call like ``update_from_suffix(model.unknown_parameters, vals)``.
+    values : iterable of numbers
+        New numerical values for the components referenced by the suffix.
+        Must be the same length as ``suffix_obj``.
+    """
+    # Check that the length of values matches the suffix length
+    items = list(suffix_obj.items())
+    if len(items) != len(values):
+        raise ValueError("values length does not match suffix length")
+
+    # Iterate through the items in the suffix and update their values
+    # Note: items are tuples of (component, suffix_value)
+    for (comp, _), new_val in zip(items, values):
+
+        # update the component value
+        # Check if the component is a VarData or ParamData
+        if isinstance(comp, (VarData, ParamData)):
+            comp.set_value(new_val)
+        else:
+            raise TypeError(
+                f"Unsupported component type {type(comp)}; expected VarData or ParamData."
+            )
